Allower should pay for the transferFrom fee
Reason
The previous implementation deducted the transferFrom fee from the
initiator of the transferFrom method which assumed the initiator has a
token balance greater than or equal than the fee. This was an
unnecessary assumption.

Implementation
1. when Alice approve B 100 coin, internally it is maintained as if she
approved 100 + fee coins
2. when Bob initiates the transfer, his allowance is decreased by 100 +
fee, and receives 100 coins. This is equivalent of Alice paying the fee
on transferFrom.